workflow: support custom TS version for playground (#8735)
